dashboard_gui/web_gui/requirements.txt
opencv-python<4.10,>=4.8.0
Pillow>=10.0.0
matplotlib>=3.7.0
numpy<2.0,>=1.24.0
psycopg2-binary>=2.9.0
Flask>=2.3.0
flask-socketio>=5.3.0
python-socketio>=5.8.0
paho-mqtt>=1.6.0
# cv-bridge is typically installed with ROS2, not via pip

